pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-spectral-rewiring-for-heterogeneous-molecular-graphs"
version = "0.1.0"
description = "Adaptive spectral rewiring for molecular graph neural networks to address over-squashing"
authors = [{name = "Alireza Shojaei"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "dgl>=1.1.0",
    "networkx>=3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "rdkit>=2023.3.1",
    "ogb>=1.3.6",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "mlflow>=2.8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
